fix: Member, MemberMatching API 분리
